requirements.txt
streamlit>=1.31
tensorflow>=2.9
Pillow>=10.0
numpy>=1.23
requests>=2.31